updating documents and prep for release
