+++ Soft_Reset_Bots/Executables/README_Instructions.txt
@@ -1,0 +1,9 @@
+Universal SR Bot INSTRUCTIONS
+
+1. Boot your emulator and load your game
+2. Encounter your target pokemon
+3. Pick a spot where the color pallet should change when shiny
+4. Launch the bot and immediately put your mouse on the spot you picked
+5. As soon as the exe loads SR your game and re encounter the target Pokemon
+6. When you get back to the position where your spot is visible press the esc button
+7. If done correctly the bot should start soft resetting your game and will stop when it finds a shiny
