+++ Soft_Reset_Bots/WIP/send_input.py
@@ -1,0 +1,83 @@
+# http://www.gamespp.com/directx/directInputKeyboardScanCodes.html
+
+import ctypes
+import time
+import pyautogui
+import win32gui
+
+SendInput = ctypes.windll.user32.SendInput
+
+W = 0x11
+A = 0x1E
+S = 0x1F
+D = 0x20
+Z = 0x2C
+UP = 0xC8
+DOWN = 0xD0
+LEFT = 0xCB
+RIGHT = 0xCD
+ENTER = 0x1C
+
+# C struct redefinitions
+PUL = ctypes.POINTER(ctypes.c_ulong)
+class KeyBdInput(ctypes.Structure):
+    _fields_ = [("wVk", ctypes.c_ushort),
+                ("wScan", ctypes.c_ushort),
+                ("dwFlags", ctypes.c_ulong),
+                ("time", ctypes.c_ulong),
+                ("dwExtraInfo", PUL)]
+
+class HardwareInput(ctypes.Structure):
+    _fields_ = [("uMsg", ctypes.c_ulong),
+                ("wParamL", ctypes.c_short),
+                ("wParamH", ctypes.c_ushort)]
+
+class MouseInput(ctypes.Structure):
+    _fields_ = [("dx", ctypes.c_long),
+                ("dy", ctypes.c_long),
+                ("mouseData", ctypes.c_ulong),
+                ("dwFlags", ctypes.c_ulong),
+                ("time",ctypes.c_ulong),
+                ("dwExtraInfo", PUL)]
+
+class Input_I(ctypes.Union):
+    _fields_ = [("ki", KeyBdInput),
+                 ("mi", MouseInput),
+                 ("hi", HardwareInput)]
+
+class Input(ctypes.Structure):
+    _fields_ = [("type", ctypes.c_ulong),
+                ("ii", Input_I)]
+
+# Actuals Functions
+
+def pressKey(hexKeyCode):
+    extra = ctypes.c_ulong(0)
+    ii_ = Input_I()
+    ii_.ki = KeyBdInput( 0, hexKeyCode, 0x0008, 0, ctypes.pointer(extra) )
+    x = Input( ctypes.c_ulong(1), ii_ )
+    ctypes.windll.user32.SendInput(1, ctypes.pointer(x), ctypes.sizeof(x))
+
+def releaseKey(hexKeyCode):
+    extra = ctypes.c_ulong(0)
+    ii_ = Input_I()
+    ii_.ki = KeyBdInput( 0, hexKeyCode, 0x0008 | 0x0002, 0,
+ctypes.pointer(extra) )
+    x = Input( ctypes.c_ulong(1), ii_ )
+    ctypes.windll.user32.SendInput(1, ctypes.pointer(x), ctypes.sizeof(x))
+
+def buttonPress():
+    handle = win32gui.GetForegroundWindow()
+    SendMessage = ctypes.windll.user32.SendMessageW
+    SendMessage(handle, 256, 0x58, 0)
+    time.sleep(1)
+    SendMessage(handle, 257, 0x58, 0)
+
+if __name__ == '__main__':
+    pyautogui.click()
+    for i in range(10):
+        buttonPress()
+        # pressKey(Z)
+        time.sleep(1)
+        # releaseKey(Z)
+        # time.sleep(1)
